\docType{methods}
\name{show}
\alias{show}
\alias{show,Dasst-method}
\title{Show method for class \code{\linkS4class{Dasst}}}
\description{
  \code{show} shows a few contents of an object of class
  \code{\linkS4class{Dasst}}.
}
\details{
  This method shows the contents of the first table stored
  in an object of class \code{\linkS4class{Dasst}}. It
  displays values limited to a few records. Use the
  \code{\link{print}} function for more options.
}
\examples{
data(plantGrowth)
plantGrowth
}

