\name{Dasst}
\alias{Dasst}
\title{A constructor for an empty object of class Dasst}
\usage{
  Dasst()
}
\value{
  An empty object of class \code{\linkS4class{Dasst}}.
}
\description{
  \code{Dasst} is a constructor for the user that returns
  an empty object of class \code{\linkS4class{Dasst}}.
}
\details{
  This function constructs an empty object of class
  \code{\linkS4class{Dasst}}. No arguments are required.
}
\examples{
myObj <- Dasst()
myObj
class(myObj)
length(myObj)
}
\seealso{
  \code{\linkS4class{Dasst}} for class definition.
}

